Feat: Added datetime code to convert data type and make plots
Changed the 'start_time' column from string to datetime object. Using
this so we can make a plot of usage by the hour over a day, by the day
across a week, and by month over the year. Also reorganized much of the
file for future plots. Moved old code to bottom cells for reference.